[Bugfix] Use StringParameter.valueFromLookup to resolve target cluster endpoint in Fetch Migration stack
This is required instead of StringParameter.valueForStringParameter because the latter returns a token that is only resolved at deployment time. Since the value of the target cluster endpoint is substituted into the Data Prepper pipeline file prior to the deployment, it incorrectly wrote a string representation of the Token object - "${Token[TOKEN.123]}". Using StringParameter.valueFromLookup looks up the value immediately, so now the substitution occurs correctly.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>