WIP to finish setting up various replayer transformation plugin projects.
I've added a TransformationLoader helper object to replace buildlDefaultJsonTransformer() for the TrafficReplayer.  This loads transformation plugins via ServiceLoader and configures them.  The configuration part still has work to be done on it.  TrafficReplayer has a new parameter to pass a configuration string, which is expected to be JSON.  When there's only one transformerProvider, the string is optional.  When there's more than one that's been loaded, the string is mandatory so that the loader knows what order to run the transformations in.  It turns out, that we can probably build very complex transformations dynamically just with configurations (e.g., here's 3 transformations that use a JSON manipulation script like jolt or JMESPath and each can be configured with their own script).

I've moved test support code for the trafficReplayer from test to testFixtures.  That gives better clarity and also allows those to be reused from other projects.  I'm not sure whether it's better to run full transformer tests from the plugin projects or the replayer.  The replayer won't be aware of every future, hypothetical transformer plugin.  Providing a more testable interface utilizing replayer components will give those components the ability to do more complex tests which are completely at the discretion of the plugin author.  What I don't like about this approach right now is the dependency list for the testFixture.  It includes netty, jackson, and junit5.  Considering how terse and lightweight the contract is for a transformer, those could be heavy and potentially conflicting assumptions.

There is still one failing test: org.opensearch.migrations.replay.HeaderTransformerTest#testMalformedPayload_andTypeMappingUri_IsPassedThrough

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>