[Fetch Migration] Suppress exit code for graceful termination (SIGTERM)
The orchestrator now suppresses exit code 143 if it is returned by the migration monitor, since this is a result of a workaround for https://github.com/opensearch-project/data-prepper/issues/3141 .
Unit tests have been added to cover this scenario. This change also ensures that a return_code is always returned by the function, and updates some stale comments.

Signed-off-by: Kartik Ganesh <gkart@amazon.com>