Remove all instances of MetricsLogger and the OpenSearch analytics domain.
Strip opensearch analytics from the docker solution and the CDK.
Use the AWS Distro for Otel's collector as the base image for our otel collector instead of the custom built one.
Kill the OpenSearch analytics domain and its bastion host that came along with the otel-collector, leaving only the collector (which will shoot metrics and traces to AWS).
Updated the README and have made the cdk.context.json set otelCollectorEnabled = true, though maybe the default should be true & this should be opt-out.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>