[Feature] Add Basic Support for OpenSearch Ingestion controls from the Migration Console (#621)
This change introduces basic support for creating/starting/stopping an OpenSearch Ingestion pipeline.

It provides an experimental CDK option for enabling OSI on the migration console. This option does not prevent creating the OSI python tool on the Migration Console container, but rather controls creating the IAM roles and policies related to OSI

It sets up necessary IAM policies for the Migration Console to assume as a 'Manager" of the OSI pipeline. It also creates a pipeline role that will be assumed by OSI to communicate with the source and target clusters

Adds the osiMigration.py script to the migration console for controlling OSI pipelines. This script supports starting/stopping a pipeline as well as creating a pipeline from provided command arguments or creating the pipeline from inferred settings based on the associated deployment.

See README update for details on executing. Once this is merged, I expect to copy and paste something similar to the wiki.

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>