Add job that passed when all other CI jobs have completed
GitHub allows 'requiring' a job to be run before a pull request can be
merged, by adding a new job that 'needs' all the other ci jobs, this job
will run last and only if they all pass.

After this is merged someone from the admin team can update the repo
settings to enforce this requirement.

Signed-off-by: Peter Nied <peternied@hotmail.com>