RFS Work Coordinator logs errors now (#837)
* RFS Work Coordinator logs errors now

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Added unit tests for the RFS OpenSearchWorkCoodinator

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Streamline test case

Signed-off-by: Peter Nied <peternied@hotmail.com>

* Updates per PR Comments

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Deleted a rogue comment

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Make Spotless happier

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Adding a new test to affirm the logging of unexpected responses for RFS Work Coordination, I noticed an API bug in the Http Client/Responses.
There was too much wiggle room between the bytes and stream payload accessors.  The logging code assumed that the stream would be recoverable, but as per the existing interface, that wouldn't always be the case (and wasn't in the way that I coded up the first test case that I had tweaked).  Now there's only one API, which requires all payloads to be recoverable indefinitely, mooting the bug.
Also, the existing tests were made a big tighter/clearer and a new test was added that goes through the public APIs to make sure that those also log (in case getResult ever gets dropped from their codepaths).

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>

* More updates per PR comments

Signed-off-by: Chris Helma <chelma+github@amazon.com>

* Fixed a typo

Signed-off-by: Chris Helma <chelma+github@amazon.com>

---------

Signed-off-by: Chris Helma <chelma+github@amazon.com>
Signed-off-by: Peter Nied <peternied@hotmail.com>
Signed-off-by: Greg Schohn <greg.schohn@gmail.com>
Co-authored-by: Peter Nied <peternied@hotmail.com>
Co-authored-by: Greg Schohn <greg.schohn@gmail.com>