Bugfix.
For the OpenSearchWorkCoordinator, acquiring leases is a two-step process.  The first step is to run an update and if that was successful, the second is to find out which work item has been assigned (this is the best that could be setup given semantics of Elasticsearch/OpenSearch).  If the query call fails, it will throw an exception and kill the process.  We won’t try again and the lease will sit in an unclaimable state for any other potential worker until it expires.  Making matters worse, the lease duration will have doubled over the last time, so every time that this happens, it’s an even more expensive error.

This fix throws a couple different errors from the query call (getAssignedWorkItem and will now retry them.  The malformed document exception isn't retried indefinitely since that's likely a severe programmer bug.  The other errors are considered transient and are retried for the duration of the lease duration (initial one, since we may never know what the lease was actually set to).

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>
Signed-off-by: Arun Lakshmanan <arnlaksh@amazon.com>