Reduce the sources for ParallelDocumentMigrationsTest
After looking into this test its duplicating much of the functionality
of EndToEndTest in the same project but the source clusters are
functionality not that different when they are flattened into a
snapshot.

The startup cost for these tests is ~2min and then ~30sec a total of
~4.5min per source.  I'm removing the test I added and pairing down to
ES 7.10 to keep our CI jumps running faster.

Signed-off-by: Peter Nied <peternied@hotmail.com>