Specify User Agent for Console CLI requests (#952)
Allows setting a global user agent on CDK deployments that will be configured in a client_options section in the services.yaml. This user agent will currently be provided to the Traffic Replayer for its replayed requests (existing logic) and to the Console CLI for all of its boto3 requests and python requests library requests. This can also be extended to other migrations as support is added.

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>