Move original captureProxy chart from k8s and fold the k8 version into this too.  This commit also creates general purpose helper scripts.
Helper scripts for the chart to handle threading command line params through ConfigMaps and into the command line.  This takes parameters from the values file and put them into ConfigMaps, always creating an immutable one for a default and possibly another for the user to override the setting at runtime.  That runtime configmap is created with no value.

The pod itself pulls the configmap values in as env variables, preferring the overridden one if it was present.  It then strings all of those into the command line invocation.
Notice that there are 3 types of values for parameters - ones with arity 0, arity 1, and arity > 2.  For arity 0, the keyname will be 'present', for arity 1, the keyname=value, and for greater arities, the data should be passed as a yaml list via the keyname 'list'.  For lists, we'll use yq and xargs to parse the contents of the yaml list to turn that into the command line - so those two utilities need to be present in the image that we're running in to take advantage of that feature.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>