Add analysis workflows for SonarQube and CodeQL
This repo has an asychronous process where SonarQube rules are run
several hours after code has merged.  This creates a life-cycle where
code is reviewed approved and merged and then needs to be updated in a
second PR.

This change introduces two new workflows to minimize the afterwork
required for changes in this repo, by running CodeQL which is built into
GitHub for security risk detection and SonarQube with its set of issues
based on security issues, code smells, and bug detectors.

These new workflows are being added without any fixes or adjustments to
the codebase to be handled in an ongoing basis.

Signed-off-by: Peter Nied <peternied@hotmail.com>