Further work to support bulk delete, index create with type mapping migration, and regex target index patterns. Fix regex replace to use java's regex classes instead of the builtin ones.  The builtin filter uses Google's re2j library, which doesn't do backreferences.  Now both capture and replace use custom filters that use the builtin java regex library.  Tests are in place to do some index remapping w/ backreferenced captures. For bulk APIs, only the delete command is fully supported, but the structure for the others should be in place. I've also worked on improving the test for creating an index and fixed numerous issues there.
It's also CRITICAL to note that constructing dictionaries in jinjava with keys that are defined via variables is NOT SUPPORTED.  See https://github.com/HubSpot/jinjava/issues/379.  The workaround that I'm currently using is to construct a map as a string in json and parse it into a map - or to construct maps dynamically.  Also note that python dictionary operations like 'update' or 'delete' are NOT present since we don't have the luxury of overriding the dictionary implementation easily.  This is also something that has a significant impact on the ease of use for jinjava and to maintain compatibility to the python implementation.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>