Fix memory leaky est
Checking allocations on the test, previously it was using between
400-480mb of memory while running, add additional jars for lucene7/9 and
we are very close to running out of memory constantly.  I've dialed down
the document to 10mb - still chunky, but with all the testing and
duplication representations the test now takes up ~180mb of memory so we
should be clear.

Signed-off-by: Peter Nied <peternied@hotmail.com>