Fix a broken jenkins pipeline where the groovy file had a syntax error... (#1255)
Plus add a new ability to test those configurations before we push them out.  I'm not really satisfied with the amount of effort/change to put a test in place that doesn't do much locally, considering that the pipeline DOES fail when there was a configuration issue.  Others might find some more value in it though.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>