POC Metadata migration support for ES2 indices
This is a proof of concept for writing a JavaScript transformation that will update indices and index templates that are using the string type, see `docs/breaking-changes/StringMappingType.md` included in this PR for additional details.

MultiTypeMappingTransformationTest.java is augmented to use the new transformation.  Run this scenario yourself with the following command.  It will 'minify' [1] the javascript into the json definition, then start the test that will load use that configuration.

```bash
python MetadataMigration/src/test/resources/updatejson.py && ./gradlew MetadataMigration:isolatedTest --tests *MultiTypeMappingTransformationTest.migrateFromUpgrade`
```

Signed-off-by: Peter Nied <peternied@hotmail.com>