Further work on an RFS pipeline.  There’s a lot in flux on how to parse configurations between the helm template of the workflow template; between argo steps; & argo and the k8s manifest of the RFS replicaset.  The idea is to use a json configuration with NO GLOBALS for the RFS settings.  Instead, the caller specifies a list of completely materialized settings that are each passed to an RFS pipeline.  That lets a user easily see what went into a workflow and we don’t have to worry about shared state.
Each RFS pipeline (“run-bulk-load”) takes in a json configuration, merges it with default values (and default empty values are important, because argo’s ‘jsonpath’ function doesn’t do well when a key isn’t found at all).

Then the next steps (attempt to) run RFS as a replicaset, while checking if RFS was done (totally mocked out code for now).  Lastly, it should delete the replicaset initially setup.  Notice that the ‘name’ value is a unique token appended to the replicaset and should be used to scope everything about each of the bulk loaders so that we can run multiple variants in parallel.  That’s still in-progress since RFS itself deals with a global coordination index.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>