Add ES 5.x K8s Local Jenkins Pipeline (#1382)
This changes introduces the test automation library to handle the operations required for installing helm dependencies, deploying our MA helm chart (along with source and target test clusters), monitoring the k8s cluster state, running our python E2E tests against the deployed environment, and displaying the results in a pleasant format.

With this library in place, a new Jenkins pipeline was added to execute our python E2E ES 5.x tests. With most of the interesting logic happening in the library, this pipeline script should be pretty thin, and be able to be reused for similar pipelines in the future.

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>