Fix soft delete behavior with ES 6 and above
Previous Soft delete logic was hiding the segment name from resolving as it is wrapped in a FilterCodecReader which needed to be unwound.
Also, assumed soft deletes were added in ES 6 when they were actually added in ES 6.5 as an opt-in feature. This miss meant we were previously recreating soft-deleted docs on the target (target could contain more docs than valid docs on the source)

Signed-off-by: Andre Kurait <akurait@amazon.com>