I've got a snapshot being created in an S3/Localstack bucket ... and the snapshot creation code workflow is returning a json block that describes the snapshot for future phases.  To get those snapshots working, parameter parsing was tweaked extensively.  Snapshot steps now rely upon the migration console to run `snapshot create` and `snapshot status`.  Notice that there's currently a CRITICAL bug in the migration console that it isn't preserving status codes, so it's not possible to kill a workflow immediately because of an errant step.  This will HAVE to be fixed, but its separate from the workflows themselves (at least I expect that the workflow steps are honoring entrypoint exit codes and propagating them).
The create-snapshot step first DELETES any existing snapshot. This seemed more prudent to guarantee a fresh snapshot.  If we want to preserve the old snapshot and not run a new one, I'm planning to manage that outside - the whole step (and branch) shouldn't likely run in that state.

All filesystem snapshot configuration for RFS has been removed.  I didn't want to support two different environments.  Local FS snapshots are great for unit tests and other places, but where we have object storage, we should really make the object storage experience great first.

The testClusters now uses a searchguard 7.10 image.  The configuration hardcodes all of the searchguard settings that were in the image (by copying, yuck) and hardcodes the localstack URI for the S3 plugin that's now included in that 7.10 searchguard base image.  Also for the base image, "test" is added to the keystore as the AWS keys.  Since that will always work for localstack and the image is only for testing, this didn't seem unreasonable (others can overwrite this with more appropriate credentials as they wish, though this image shouldn't be used for production anyway).  Putting those within the docker image simplified bringing the container up.

How the migration console is utilized has changed.  I don't try to write a volume, but instead cons the services yaml configuration, then simply pass that as a parameter to a container that writes the services yaml file, then runs a specified command (that also came in as a parameter).  That simplifies a lot.  Ephemeral volumes don't work because they're bound to the lifetime of a pod.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>