Refactor hard-coded snapshot repo name to be provided (#1472)
This change refactors out the hardcoded snapshot repository name found in several places in our code base and allows passing this name. This allows our migration console to be more customizable when it comes to creating a snapshot or performing operations such as checking the status of an external snapshot.

Also included is some refactoring of some disjoint snapshot logic in the console library to be contained within the snapshot class
---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>