Improve Kafka export and loading workflow (#1481)
This change fixes the currently broken kafka export script and improves it to store both the original connection id and the actual kafka record value.

This also adds a small utility which will take this kafka export gzip file as an argument and load the contents of the export script into Kafka via a Kafka broker endpoint and using the Kafka Producer pattern. Also included are instructions for taking a Kafka export file and loading it into a local K8s environment.

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>