Add experimental path for EKS solutions deployments (#1546)
This change adds an additional experimental template for our solutions CDK to generate a CFN stack that utilizes an EKS cluster. I expect this to change a good bit as we discover more about how we should configure EKS and what additional resources we actually need there, but should provide developers of this feature a good starting point.

The template can be generated with the following command:

cdk synth "Migration-Assistant-Infra-Create-VPC-v3"

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>