Fix TrafficReplayer behavior with HEAD and CONNECT method requests.
TrafficReplayer's HttpByteBufFormatter, used for tuples, decodes the request and response in separate Embedded Channels. This limits the ability for the decoder to detect rfc9112 6.3.1-2 behavior where responses from HEAD and CONNECT requests are terminated after the headers ignoring content-length fields.

We now achieve correct behavior for these requests by writing a LastHttpContent.EMPTY_LAST_CONTENT in the channel if the output was empty after the entire request/response was sent. This triggers the end of the message and output is written.

This change converts from using HttpClientCodec to HttpResponseDecoder for initial response decoding and HttpServerCodec to HttpRequestDecoder for request decoding to simplify the decoding and make clearer the responsibilities. HttpClientCodec misleadingly handled this case for HEAD and CONNECT requests as it seeks to maintain state for these responses, but to do so, it requires the request be sent through the outbound of the channel. Doing this would add unneccissary complexity to couple the request and response decoding for the HttpByteBufFormatter especially in cases of multiple responses during retrys where we would need to decode a single request several times to maintain this state for each of the response parsing.

Signed-off-by: Andre Kurait <akurait@amazon.com>