Allow building images to any Docker registry and image builds in containers (#1545)
This change introduces a new pattern for building images to our tooling. It makes use of Jib for our Java applications (except for RFS currently as it has additional Dockerfile steps) and BuildKit for our other images, to allow building images to any specified docker registry, whether that be a local registry or a remote registry like AWS ECR. This pattern also allows for building the images with a container, which the bootstrapK8s helm chart utilizes to be able to pull this Github repository and build images within a K8s pod.

At a top level this can be executed, after following the setup steps in buildImages.md, like so:

./gradlew buildImagesToRegistry
And includes options for specifying the registry endpoint as well as the architecture to build with:

./gradlew buildImagesToRegistry -PregistryEndpoint=123456789012.dkr.ecr.us-west-2.amazonaws.com/my-ecr-repo -PimageArch=arm64

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>
Co-authored-by: Peter Nied <peternied@hotmail.com>