Reduce default RestClient connection and request timeout (#1746)
We have seen some worker coordination client/server time out of sync. Upon investigation
this could occur during connection timeout issues since a reqeust could be delayed to be sent
until after a connection timesout (prior default 30s) which would cause a clock skew exception
with the current value of 15 seconds. This change sets the value to 5 and further failure will delegate to
the retry logic surrounding a call which would be able to reconstruct the request as needed.

Signed-off-by: Andre Kurait <akurait@amazon.com>