feat(migration-assitant-solution): fetch latest AMI from SSM parameter store
This is more of a refactoring. During a recent change, I realised the AMI IDs
referenced (generated by the `create-ami-map.sh`) were quite outdated.
(ie. The image was deprecated)

As a solution, I refactored this code to use SSM Parameter store to fetch
respective AMI ID. As those parameters are public, and contains the current
AMI IDs, this will always fetch the latest AMI ID.

This change will also remove the need for the `create-ami-map.sh` script.

These are documented in the AL2023 EC2 documentation as well as the SSM parameter store one:
- https://docs.aws.amazon.com/linux/al2023/ug/ec2.html
- https://docs.aws.amazon.com/systems-manager/latest/userguide/parameter-store-public-parameters-ami.html

Signed-off-by: Mert Akengin <5755154+spacelatte@users.noreply.github.com>