AWaaIT - Argo Workflows as an Integration Test (#1770)
Restructures the current ma_workflow_test.py integration tests to use the migrationAssistantWithArgo Helm chart which utilizes Argo workflows as its means for performing migrations. The fullMigrationWithClusters Argo workflow template these tests utilize currently supports Metadata and Backfill migrations, and plan to be extended to Live Capture migrations as support is added into the fullMigration workflow template that it utilizes. This workflow template allows the workflow to create the source and target clusters utilized by the test, giving the testing structure more flexibility with defining clusters as well as reducing the overhead to execute these tests. The workflow is currently suspended at two points, after creating the clusters and after the migration is complete, to allow loading sample data and verifying the clusters after migration respectively, but seemingly could be expanded to more if it seems appropriate.

Some useful benefits that come from this pattern include:

Each test case is given a fresh set of clusters to use, that the test case defines
Multiple workflows can be ran in parallel as each workflow is using its own clusters and migration resources
Application parameters such as transformation config, can be defined by the test and do not need to be provided beforehand


---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>