Improve Developer UX with Argo Integration Tests (#1797)
Focus of this change was adding the ability to reuse clusters in the Argo integ test suite that is triggered by the testAutomation library. I went through a few iterations of adding this functionality with the least complexity. I am not pleased with some of the conditional logic the `MATestBase` currently has but preferred this for the time being after seeing some of the complexity/difficulty of trying to weave this conditional logic into the workflow file itself (I'm hopeful the Typescript DSL would make this much easier). A next iteration here could benefit from having a workflow step that calls some python logic to see if a cluster exists or otherwise creates the cluster, and always outputs the config to make things simpler.

As noted in the README, a command like the following, could be used when developing to reuse clusters across multiple executions
```shell
pipenv run app --dev --source-version=ES_5.6 --target-version=OS_2.19 --test-ids=0001
```

Additional changes this PR adds
* Modifies Argo Helm chart to enable the API server for the migration console
* Adds ability to enable migration console developer mode automatically when running testAutomation
* Adds ES_7.10 as another source for the integ tests
* Adds OS_1.3 as another target for the integ tests

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>