Update Kafka version to 3.9.1 (#1817)
This updates the Kafka version used across the project to 3.9.1. The MSK instance in CDK is changed to 3_9_X and the kafka clients are already on 4.0.0 and did not need update. By using a newer non-archived version of Kafka we can also utilize the CDN for the kafka distribution download in the Migration Console. Since this version could also later be added to the archived versions, there is a fallback URL to use the archive if the CDN endpoint is not available.

---------

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>