Bugfixes around type correctness for expressions and how Argo interprets them
The '=' after `{{` coerces all json output to be encoded (with escapes) rather than just dropped as-is, so I've gotten much more careful about when to use it.  I'm sure that there will still be more issues around this, though they may eventually be solved by having multiple expressions within a templated string rather than trying to keep everything in one expression.

PlainObject now includes a Serialized<> option.  This represents a string that when deserialized will return the object type that it represented.  This type is used by jsonPath operations.  The sprig function works on strings rather than objects - but by keeping the type around, we can still get perfect type checking and autocompletion.

Making the change to allow Serialized included coercing the input parameters through that so that every usage of the input parameter as an expression would be understood to be serialized, not an object.  That was always a major source of confusion when dealing with Argo's YAML representation directly.  Now it's much more clear and type-safe.

To keep the ergonomics, the register() functions set up their fields so that they can accept Serialized<T> objects if the input types were already T (in addition to literal types of T),  That means that one can still route inputs directly into downstream template calls

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>