Making some expressions safer & getting what appears to be valid services.yaml contents.
The services.yaml generation took a lot of trial and error, but eventually made a lot of sense.
In the old YAML files, it was a bit sloppy.  Here there are a few asString casts, but at least it starts to show when we need to parse as json and serialize as it.
recordToString now returns a BaseExpression intersection so that it can act as a string or record type, but I haven't figured out how to unlock that yet - if I would have that wired correctly, I wouldn't have needed as many asString conversions.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>