Add a the 'dig' function to (soon) replace jsonPathLoose, plus more.
jsonPathLoose doesn't work because it throws when a value isn't present.  There's no way to push a default value in.  There's no way to catch the panic.  I tried a ton of different things thrown out of LLMs & I've seen nothing in the docs around exceptions.  There's default functions to choose a value over null, but nothing for when there was an error.
Anyway, the dig function works ergonomically very much like the jsonPath functions, letting the user get autocompletion at each level.  Unlike jsonPath, it accepts a record instead of a serialized record.

I've also introduced an infix operator that replaces the arithmetic operator.  Add and or are now available.  Side note - I think that like helm (which uses gotemplate) - there's no fast failing on conditionals, which makes the error throwing thing even more difficult to workaround.

Expression formatting has a major enhancement.  I was having difficulty getting "{{=input.parameter.alreadyNamed == ''}}" and many variants to work.  When conditions are already in an expression context, so the opening "{{=" is redundant.  The formatExpression function now takes a boolean in to determine if it should use markers for identifiers and we use that mode for conditions.  Now, the previous example is rendered as `{{inputs.parameters.alreadyDefinedName}} == ''`

Other expressions changes are
 * deserializeRecord which is a convenience function/alias for stringToRecord that doesn't require a typeToken, deriving the type from the expression's serialized type.
 * toLowerCase, toUpperCase
 * dig, and, or (previously mentioned)
 * nullCoalesce is still present, but an old implementation that didn't use a FunctionExpression was removed

With the current createMigration.yaml & the testClusters helm chart installed, I can now create a snapshot and make it into the metadata migration container (which then crashes because of arguments of something like that)

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>