Fix a broken when check.
When a value is computed to be the empty string, that results in argo value coercion that makes the input parameter null.  Therefore, testing nil == '' is ill-formed and won't work.
Instead, in this case, I just had to cast it to a boolean, C-style.  I'd like to think about how to bake this into the DSL codification.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>