Lots of cleanup & progress.
Increased the type strength of the dig command so that the return value and default value are narrowly scoped.  To do that, I had to make the path land inside an array so that it can be disambiguated from the default value. It's a bit clunky, but it's typesafe, which is super critical to have.
Put single quotes around parameters substituted during the "Internal" expression renderings (for when conditions).  Skpping some of the nodes now works properly (& inversely, running them when they should be run!)
When conditions can now be of two varieties - either argo expressions, which are expected to just return the word "true" or "false" or govaluate ones.  The former will be run and THEN the govaluate expression will evaluate (so true/false are value govaluate expressions already).  That's why we need single values around inputs, there are two passes of parsing.
Remove the empty configmap in lieu of a very long doc comment for setupLog4jConfigForContainer.  I also noticed that the volumes weren't being set up for commonWorkflowTypes' utilities, so now that's being done.
Lastly, I've renamed the script to make the schemas to makeTemplates & the npm command is `npm run make-templates`

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>