Add a couple other functions and automatically serialize default values for dig...
On the dig change, I'm not positive that that's the right thing to do.  I'm currently testing out argo behavior and may further adjust.
I think that the serialization and widening changes that it forced were a good idea though, so those should stay in either way.
I've made some other minor adjustments too, like making OPTIONS optional in the user schema and replacing jsonPathStrict for one of them with dig

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>