WIP - enable environment variable changes and splitting metadata migration into evaluate & migrate.
To match the changes to parse inputs from environment variables, the env variables for Argo now are constructed with a prefix and a suffix on each variable.
The MetadataMigrate template had been using the migration console templates.  Rather than extend the services.yaml to pick up the metadata configuration, I chose to make the metadata tasks just run the java command directly rather than invoking it through the console.  That provides much more direct control and is much easier to maintain.
The MetadataMigration command itself has a positional argument ("evaluate"|"migrate") that isn't modeled as a Parameter.  Passing that as an environment variable would have required some significant changes to the java code.  Since all of the other arguments are identical for each of the modes, I didn't want to have two templates.  To accommodate the new value, I pass it as an argo input now from the two different steps that call the same template.  The command now does an argo replacement.  Previously command took a string literal.  NOTICE that the ARGO command cannot take a BaseExpression<string[]> but it can take BaseExpression<string>[] - it runs the template substitution on each individual array element.
I've added some other convenience functions for env variable setup to be able to pull and transform from records.
METADATA_OPTIONS should now be complete and the code is beginning to pass it through (I'm still working on getting an entire run to succeed).

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>