Update solutions AMIs
We've recieved reports that the solutions project AMI need to be updatd
due to some security issues that have been discovered.  I've updated all
images to the latest release, including gov regions.

I've also changed the source for the images, previous we would use the
first image returned by searching the AMI catalog, instead we are
calling the Amazon managed parameter that holds these details for all
regions.

I've also updated the script to detect if there are missing regions that
need to be opted into since we saw a new region arrive "ap-east-2", this
should prevent folks that run this script in the future from missing any
existing or new regions.

Finally I've also sorted the generated list, which should make it
clearer to see any updates in the future.

Signed-off-by: Peter Nied <peternied@hotmail.com>